Redirect aleph permalink pattern to Primo
Why are these changes being introduced:

* Graceful degration of permalinks during system migration will be
  valuable to support both users with bookmarks, search engines, and
  web pages that cannot update to new links until after the new system
  goes live because the new permalinks are not yet known.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-2110

How does this address that need:

* This uses Alma SRU queries to look up an Aleph ID, obtain the Alma ID
and then contruct a URL to Primo UI.

Document any side effects to this change:

* It's possible redirecting directly to Primo UI is not the desired
solution. This work will still be useful if that is the case, as the URL
we generate in this would need to be passed to the generic transition
page if direct redirects are not desired.

* Future work to update old aleph permalinks on our internal web pages
is necessary before we can retire this system.